// src/app/page/master/jenis-deskripsi/jenis-deskripsi.interface.ts
 import {EventEmitter} from '@angular/core';

    export interface JenisDeskripsi {
	kodeJabatan?: any;
	namaPangkat?: string;
	reportDisplay?: string;
	noUrut?: number;
	kdGolonganPegawai?: number;
	ruang?: string;
	kdPangkatHead?: number;
	KodeExternal?: any;
	kodeExternal?: any;
	namaExternal?: string;
	kdDepartemen?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
  	value?: any;
  	namaGolonganPegawai?: string;
  	id?:any;
  	kode?:any;
  	version?:any;
	namaJabatan?:any;
	jenisJabatan?:any;
	levelJabatan?:any;
	Version?:any;
	namaJenisDeskripsiHead?: string;
	namaJenisDeskripsi?: string;
    }